* Added SettingsManager * Started shifting of settings to SettingsManger * Finished NewFileDialog * Unified datatypes in libZelda and libWiiSave * Fixed a bug in libWiiSave where writeU32 was trying to swap a 32bit integer with a 64bit function.

+++ trunk/save editor/LibWiiSave/include/aes.h
@@ -1,13 +1,15 @@
 #ifndef __AES_H_
 #define __AES_H_
 
+#include <Types.hpp>
+
 #ifdef __cplusplus
 extern "C" {
 #endif
 
-void aes_encrypt(u8 *iv, const u8 *inbuf, u8 *outbuf, unsigned long long len );
-void aes_decrypt(u8 *iv, const u8 *inbuf, u8 *outbuf, unsigned long long len );
-void aes_set_key(const u8 *key );
+void aes_encrypt(Uint8 *iv, const Uint8 *inbuf, Uint8 *outbuf, Uint64 len);
+void aes_decrypt(Uint8 *iv, const Uint8 *inbuf, Uint8 *outbuf, Uint64 len);
+void aes_set_key(const Uint8 *key );
 
 #ifdef __cplusplus
 }
